feat: implementa servico de update de estudo e middleware de autenticacao(incomplete)

+++ src/useCases/(Study)/DeleteStudy/DeleteStudy_UseCase.ts
@@ -1,3 +1,4 @@
+import { NotFound } from "../../../repositories/IErrorRepository";
 import { IStudyRepository } from "../../../repositories/IStudyRepository";
 import { IDeleteStudyDTO } from "./DeleteStudy_DTO";
 
@@ -6,7 +7,12 @@ export class DeleteStudyUseCase {
         private studiesRepository: IStudyRepository
     ) {}
     
-    async execute(data: IDeleteStudyDTO): Promise<void | string> {
+    async execute(data: IDeleteStudyDTO): Promise<void> {
+        const studyExists = await this.studiesRepository.findById(data.id);
+        if (!studyExists) {
+            throw new NotFound("Estudo não encontrado no sistema");
+        }
+        
         return await this.studiesRepository.deleteById(data.id);
     }
 }
